refactor(CytoscapeRenderer): move export and layout into separate files
